transferring a request to app/queries
